Echo all commands
I'm trying to reason through why `make check-inter-api-level` isn't
being executed, only to discover that all commands that it would
execute are *not* being echo'd, due to the `@` prefix.

Remove the `@`, so that commands are fully echo'd, thus allowing me to
better determine what is actually happening from log files.